Scormfile model refactoring. Create Scormfiles without creating zip files

+++ app/models/zipfile.rb
@@ -5,27 +5,28 @@ class Zipfile < Document
   end
 
   def fileType
-    fileType = "Zipfile"
-    fileType = Zipfile.fileType(self.file.path) if self.file.class == Paperclip::Attachment and !self.file.path.blank?
-    return fileType
+    Zipfile.fileType(self.file.path)
   end
 
   def self.fileType(filePath)
-    Zip::File.open(filePath) do |zip|
-        manifest = zip.entries.select{|e| e.name == "imsmanifest.xml"}.first
-        if manifest
-          schema = Zipfile.getSchemaFromXmlManifest(Nokogiri::XML(manifest.get_input_stream.read)) rescue "invalid schema"
-          case schema
-          when "IMS Content"
-            fileType = "Imscpfile"
-          when "ADL SCORM", nil
-            fileType = "Scormfile"
+    if File.exists?(filePath)
+      Zip::File.open(filePath) do |zip|
+          manifest = zip.entries.select{|e| e.name == "imsmanifest.xml"}.first
+          if manifest
+            schema = Zipfile.getSchemaFromXmlManifest(Nokogiri::XML(manifest.get_input_stream.read)) rescue nil
+            case schema
+            when "IMS Content"
+              return "Imscpfile"
+            when "ADL SCORM", nil
+              return "Scormfile"
+            end
+          else
+            return "Webapp" if zip.entries.select{|e| e.name == "index.html"}.first
           end
-        else
-          index = zip.entries.select{|e| e.name == "index.html"}.first
-          fileType = "Webapp" if index
-        end
+          return "Zipfile"
+      end
     end
+    nil
   end
 
   def self.getSchemaFromXmlManifest(xmlManifest)
@@ -33,24 +34,6 @@ class Zipfile < Document
     schemaEl = xmlManifest.at_css('//metadata//schema')
     return schemaEl.text unless schemaEl.nil?
     nil
-  end
-
-  def getResourceAfterSave
-    case self.fileType
-    when "Scormfile"
-      resource = Scormfile.createScormfileFromZip(self)
-    when "Imscpfile"
-      # TODO
-      # resource = Imscpfile.createImscpfileFromZip(self)
-      resource = self
-    when "Webapp"
-      # TODO
-      # resource = Webapp.createWebappFromZip(self)
-      resource = self
-    else
-      resource = self
-    end
-    return resource
   end
   
 end
